Bindings: finish generic cleanup
