--- src/channels/plugins/acp-binding-sessions.ts
@@ -1,5 +1,0 @@
-export {
-  ensureConfiguredAcpBindingReady,
-  ensureConfiguredAcpBindingSession,
-  resetAcpSessionInPlace as resetConfiguredAcpBindingSessionInPlace,
-} from "../../acp/persistent-bindings.lifecycle.js";
