--- src/channels/plugins/acp-binding-types.ts
@@ -1,5 +1,0 @@
-export type {
-  AcpBindingConversation,
-  AcpBindingResolution,
-  CompiledAcpBinding,
-} from "./binding-types.js";
