+++ src/channels/plugins/stateful-target-builtins.ts
@@ -1,0 +1,13 @@
+import { acpStatefulBindingTargetDriver } from "./acp-stateful-target-driver.js";
+import {
+  registerStatefulBindingTargetDriver,
+  unregisterStatefulBindingTargetDriver,
+} from "./stateful-target-drivers.js";
+
+export function ensureStatefulTargetBuiltinsRegistered(): void {
+  registerStatefulBindingTargetDriver(acpStatefulBindingTargetDriver);
+}
+
+export function resetStatefulTargetBuiltinsForTesting(): void {
+  unregisterStatefulBindingTargetDriver(acpStatefulBindingTargetDriver.id);
+}
